requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
streamlit==1.40.2
openai==1.55.3
python-dotenv==1.0.1